test: Reuse fixture and add clean up op
